added tests

+++ src/main/java/org/example/DAO/ReizigerDAOsql.java
@@ -114,13 +114,13 @@ public class ReizigerDAOsql implements ReizigerDAO {
     }
 
     @Override
-    public Reiziger findById(int id) throws SQLException  {
+    public Reiziger findById(long id) throws SQLException  {
         Connection c = getConnection();
         String query = "SELECT * FROM reiziger WHERE reiziger_id = ?";
         pstmt = c.prepareStatement(query);
         c.setAutoCommit(false);
 
-        pstmt.setInt(1, id);
+        pstmt.setLong(1, id);
         rs = pstmt.executeQuery();
 
         if (rs.next()) {
